Initial attempt to convert .lyx to .xml
Add an xml_streamer.py that implements a simple XmlStreamer class for
streaming out XML documents.

Add a lyx2xml.py that implements a simple-minded conversion of .lyx
to .xml, where things like \begin_layout Foo -> <foo> ... </foo> and
so on.

The idea is to have a faithful LyX -> XML translation such that one
could then apply XSLs to convert to arbitrary XML (and text) formats.

The schema is implied in the structure of .lyx.

Also, it should be trivial to convert from this LyX XML schema back
to .lyx using an XSL again.  This should then allow for bi-directional
import/export from/to many XML schemas and any document formats that
can be converted to/from XML.